Add files via upload
The code for the step sensor part was written by me. It mainly gets the user's step count and connects to MQTT which can be demostrated by the dashboard.py. The other code in dashboard.py is contributed by Xu Shiqi and Li Caixi. My GitHub account is lr200611@qq.com