Dashboard construct and the animation updated
This is Xu Shiqi from team2 class 2. The code for the sleep sensor was written by me it simulates the recording and transmittting of the sleep duration. I create the dashboard and the animation for the whole project. This gif of the beautiful owl is also my offering. My GitHub account is 2114599085@qq.com